third_party/blink/web_tests/svg/custom/transform-ignore-after-invalid-expected.txt

CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "…e(2,2) scale(1,2,3) translate(50…".
